Load and augment coral polyp density data

Loads polyp_density.csv, adds an estimated entry for Acropora millepora based on similar species (A. nasuta and A. spathulata), and assigns growthform classifications.

Usage

polyp_density_data

Format

An object of class data.frame with 45 rows and 7 columns.

Value

A polyp_density_data data frame with all species and assigned growthforms.

Details

Acropora millepora is not present in the original dataset.

To estimate it, data for A. nasuta and A. spathulata are duplicated and relabeled.

A new spp code "AM" is assigned to the added entries.

Growthforms are recoded using fct_recode() based on species name.

Examples

head(polyp_density_data)
#>         id spp              species cm2 polyps polyps_cm2 growthform
#> 1      764  AC    Acropora cytherea  16   1268    79.2500    Tabular
#> 2      663  AH  Acropora hyacinthus  16   1355    84.6875    Tabular
#> 3      719  AH  Acropora hyacinthus  16   1212    75.7500    Tabular
#> 4 large G.  GP Goniastrea pectinata  16     90     5.6250    Massive
#> 5      769  GP Goniastrea pectinata  16     57     3.5625    Massive
#> 6      733  GP Goniastrea pectinata  16    112     7.0000    Massive